MTBench: Moved table components and styles to the MTBench folder, deleted extra figures in assets
