+++ app/projects/mtbench/page.mdx
@@ -1,6 +1,6 @@
 
 import { Authors, Badges} from '@/components/utils'
-import { Table, Table1, Table2, Table3, Table4}from '@/components/table'
+import { Table, Table1, Table2, Table3, Table4} from './components/table.tsx'
 
 # MTBench: A Multimodal Time Series Benchmark for Temporal Reasoning and Question Answering
